edge_node/main/main/requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
certifi==2026.1.4
cffi==2.0.0
minio==7.2.20
numpy==2.2.6
opencv-python==4.13.0.90
paho-mqtt==2.1.0
psutil==7.2.2
pycparser==3.0
pycryptodome==3.23.0
python-dotenv==1.2.1
rknn-toolkit-lite2 @ file:///home/cat/rknn-toolkit-lite2/packages/rknn_toolkit_lite2-2.3.2-cp310-cp310-manylinux_2_17_aarch64.manylinux2014_aarch64.whl
ruamel.yaml==0.19.1
typing_extensions==4.15.0
urllib3==2.6.3
